+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Unreleased
+----------
+- Fixed issue when deserializing non-trade activity object without a
+  quantity
+
+
 0.9.0
 -----
 - Added `quantity` field to `NonTradeActivity` type